requirements.txt
pydantic < 2
pyyaml < 6